pyproject.toml
[project]
name = "onwaystudyapi"
version = "0.1.0"
description = "on way study api project"
authors = [
    {name = "LeandroDolensi",email = "leandro.dolensi@aluno.ifsp.edu.br"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pre-commit (>=4.3.0,<5.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
